refactor: parsing and nonstandard testing
day05 provides sample/test data differently than other days. in
particular, the sample text for part 1 doesn't really calculate the max
(like the problem does), it just verifies the seatId values for the test
data. refactor main (and supporting AocUtil logic) to support this
oddball. Along the way, expose AocUtil parse and refactor it to work in
any monadic context so that day05 can use it in Maybe. this conveniently
simplifies day04 a bit, too.